Atualização do código-fonte - Classe abstrata
