+++ limpee/src/main/java/school/sptech/limpee/domain/usuario/Prestador.java
@@ -6,85 +6,16 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.validation.constraints.Min;
 
-@Entity
 public class Prestador extends Usuario {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private final String tipoUsuario = "prestador";
-    @Min(0)
-    private int qtdServicosFinalizados;
-    @Min(3)
-    private int anosExperiencia;
-    public Prestador() {}
-
-    public Prestador(String nome, String email, String senha, String genero, int ranking, int anosExperiencia) {
-        super(nome, email, senha, genero, ranking);
-        this.qtdServicosFinalizados = 0;
-        this.anosExperiencia = anosExperiencia;
+    public Prestador() {
     }
 
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getTipoUsuario() {
-        return tipoUsuario;
-    }
-
-    public int getQtdServicosFinalizados() {
-        return qtdServicosFinalizados;
-    }
-
-    public void setQtdServicosFinalizados(int qtdServicosFinalizados) {
-        this.qtdServicosFinalizados = qtdServicosFinalizados;
-    }
-
-    public int getQtdServicoFinalizado() {
-        return qtdServicosFinalizados;
-    }
-
-    public void setQtdServico(int qtdServicosFinalizados) {
-        this.qtdServicosFinalizados = qtdServicosFinalizados;
-    }
-
-    public int getAnosExperiencia() {
-        return anosExperiencia;
-    }
-
-    public void setAnosExperiencia(int anosExperiencia) {
-        this.anosExperiencia = anosExperiencia;
-    }
-
-    public void prestarServico(){
-
+    public Prestador(String nome, String email, String senha, String genero, int ranking, String tipoUsuario, int qtdServicosSolicitados, int qtdServicosFinalizados, int anosExperiencia) {
+        super(nome, email, senha, genero, ranking, tipoUsuario, qtdServicosSolicitados, qtdServicosFinalizados, anosExperiencia);
     }
 
     @Override
     public void calcularPontuacao(Cliente c, int pontuacao) {
-        switch (pontuacao){
-            case 1:
-                c.setRanking(getRanking()-3);
-                break;
-            case 2:
-                c.setRanking(getRanking()-2);
-                break;
-            case 3:
-                c.setRanking(getRanking()+2);
-                break;
-            case 4:
-                c.setRanking(getRanking()+3);
-                break;
-            case 5:
-                c.setRanking(getRanking()+4);
-                break;
-            default:
-                c.setRanking(getRanking()-1);
-                break;
-        }
+
     }
 }
