+++ limpee/src/main/java/school/sptech/limpee/domain/usuario/Usuario.java
@@ -3,6 +3,8 @@ package school.sptech.limpee.domain.usuario;
 import jakarta.persistence.MappedSuperclass;
 import jakarta.validation.constraints.*;
 
+import javax.naming.directory.InvalidAttributeValueException;
+
 @MappedSuperclass
 public abstract class Usuario {
     @Size(min = 2, max = 100)
@@ -16,16 +18,34 @@ public abstract class Usuario {
     private String senha;
     @NotBlank
     private String genero;
+    private String tipoUsuario;
+    @Min(0)
+    private int qtdServicosSolicitados;
+    @Min(0)
+    private int qtdServicosFinalizados;
+    @Min(3)
+    private int anosExperiencia;
     private int ranking;
 
     public Usuario() {}
 
-    public Usuario(String nome, String email, String senha, String genero, int ranking) {
+    public Usuario(String nome, String email, String senha, String genero, int ranking, String tipoUsuario, int qtdServicosSolicitados, int qtdServicosFinalizados, int anosExperiencia) {
         this.nome = nome;
         this.email = email;
         this.senha = senha;
         this.genero = genero;
         this.ranking = ranking;
+        this.tipoUsuario = tipoUsuario;
+        this.qtdServicosSolicitados = qtdServicosSolicitados;
+        this.qtdServicosFinalizados = qtdServicosFinalizados;
+        this.anosExperiencia = anosExperiencia;
+    }
+
+    public abstract void calcularPontuacao(Cliente c, int pontuacao);
+
+    public double calcularMedia(int qtdServico){
+        double notaMedia;
+        return notaMedia = (ranking/qtdServico);
     }
 
     public String getNome() {
@@ -67,10 +87,39 @@ public abstract class Usuario {
     public void setRanking(int ranking) {
         this.ranking = ranking;
     }
-    public double calcularMedia(int qtdServico){
-        double notaMedia;
-        return notaMedia = (ranking/qtdServico);
-    };
 
-    public abstract void calcularPontuacao(Cliente c, int pontuacao);
+    public String getTipoUsuario() {
+        return tipoUsuario;
+    }
+
+    public void setTipoUsuario(String tipoUsuario) throws InvalidAttributeValueException {
+        if (!(tipoUsuario.equalsIgnoreCase("cliente") || tipoUsuario.equalsIgnoreCase("prestador")))
+            throw new InvalidAttributeValueException("Tipo de cliente inválido.");
+
+        this.tipoUsuario = tipoUsuario;
+    }
+
+    public int getQtdServicosSolicitados() {
+        return qtdServicosSolicitados;
+    }
+
+    public void setQtdServicosSolicitados(int qtdServicosSolicitados) {
+        this.qtdServicosSolicitados = qtdServicosSolicitados;
+    }
+
+    public int getQtdServicosFinalizados() {
+        return qtdServicosFinalizados;
+    }
+
+    public void setQtdServicosFinalizados(int qtdServicosFinalizados) {
+        this.qtdServicosFinalizados = qtdServicosFinalizados;
+    }
+
+    public int getAnosExperiencia() {
+        return anosExperiencia;
+    }
+
+    public void setAnosExperiencia(int anosExperiencia) {
+        this.anosExperiencia = anosExperiencia;
+    }
 }
