+++ limpee/src/main/resources/data.sql
@@ -1,0 +1,4 @@
+INSERT INTO
+    cliente(email, genero, nome, ranking, senha, qtd_servicos_solicitados, tipo_usuario)
+VALUES
+    ('teste@email.com', 'feminino', 'Dona Ana', 100, 'teste123', 10, 'cliente');
